pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "reportlab>=4.4.2",
    "sqlalchemy>=2.0.41",
    "pycryptodome>=3.23.0",
    "cryptography>=45.0.6",
    "bcrypt>=4.3.0",
    "flask-socketio>=5.5.1",
    "python-socketio>=5.13.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "crypto>=1.4.1",
]
